returning meaningfull request responses, refactoring for better code readability
